fix double swagger tags
