+++ services/121-service/src/registration/controllers/registrations.controller.spec.ts
@@ -2,9 +2,9 @@ import { TestBed } from '@automock/jest';
 import { HttpStatus } from '@nestjs/common';
 import { PaginateQuery } from 'nestjs-paginate';
 
+import { RegistrationsController } from '@121-service/src/registration/controllers/registrations.controller';
 import { RegistrationStatusPatchDto } from '@121-service/src/registration/dto/registration-status-patch.dto';
 import { RegistrationStatusEnum } from '@121-service/src/registration/enum/registration-status.enum';
-import { RegistrationsController } from '@121-service/src/registration/registrations.controller';
 import { RegistrationsBulkService } from '@121-service/src/registration/services/registrations-bulk.service';
 import { RegistrationsPaginationService } from '@121-service/src/registration/services/registrations-pagination.service';
 import { ScopedUserRequest } from '@121-service/src/shared/scoped-user-request';
