+++ services/121-service/src/user/controllers/user.controller.ts
@@ -10,14 +10,12 @@ import {
   ParseIntPipe,
   Patch,
   Post,
-  Put,
   Query,
   Req,
   Res,
   UseGuards,
 } from '@nestjs/common';
 import {
-  ApiBody,
   ApiOperation,
   ApiParam,
   ApiQuery,
@@ -30,33 +28,20 @@ import { THROTTLING_LIMIT_HIGH } from '@121-service/src/config';
 import { AuthenticatedUser } from '@121-service/src/guards/authenticated-user.decorator';
 import { AuthenticatedUserGuard } from '@121-service/src/guards/authenticated-user.guard';
 import { CookieNames } from '@121-service/src/shared/enum/cookie.enums';
-import { ScopedUserRequest } from '@121-service/src/shared/scoped-user-request';
-import {
-  CreateProgramAssignmentDto,
-  DeleteProgramAssignmentDto,
-  UpdateProgramAssignmentDto,
-} from '@121-service/src/user/dto/assign-aw-to-program.dto';
 import { changePasswordWithoutCurrentPasswordDto } from '@121-service/src/user/dto/change-password-without-current-password.dto';
 import { CreateUsersDto } from '@121-service/src/user/dto/create-user.dto';
-import { CreateUserRoleDto } from '@121-service/src/user/dto/create-user-role.dto';
 import { FindUserReponseDto } from '@121-service/src/user/dto/find-user-response.dto';
-import { GetUserReponseDto } from '@121-service/src/user/dto/get-user-response.dto';
 import { LoginUserDto } from '@121-service/src/user/dto/login-user.dto';
 import {
   UpdateUserDto,
   UpdateUserPasswordDto,
 } from '@121-service/src/user/dto/update-user.dto';
-import { UpdateUserRoleDto } from '@121-service/src/user/dto/update-user-role.dto';
-import {
-  AssignmentResponseDTO,
-  UserRoleResponseDTO,
-} from '@121-service/src/user/dto/userrole-response.dto';
 import { UserEntity } from '@121-service/src/user/entities/user.entity';
 import { PermissionEnum } from '@121-service/src/user/enum/permission.enum';
-import { throwIfSelfUpdate } from '@121-service/src/user/helpers/throw-if-self-update';
 import { UserRO } from '@121-service/src/user/user.interface';
 import { UserService } from '@121-service/src/user/user.service';
 
+@ApiTags('users')
 @UseGuards(AuthenticatedUserGuard)
 @Controller()
 export class UserController {
@@ -65,82 +50,7 @@ export class UserController {
     this.userService = userService;
   }
 
-  @AuthenticatedUser()
-  @ApiTags('roles')
-  @ApiOperation({ summary: 'Get all user roles' })
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description: 'Returns a list of roles and their permissions',
-    type: [UserRoleResponseDTO],
-  })
-  @Get('roles')
-  public async getUserRoles(): Promise<UserRoleResponseDTO[]> {
-    return await this.userService.getUserRoles();
-  }
-
-  @AuthenticatedUser({ isAdmin: true })
-  @ApiTags('roles')
-  @ApiOperation({ summary: 'Create new user role' })
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description: 'Returns the created role',
-    type: UserRoleResponseDTO,
-  })
-  @Post('roles')
-  public async addUserRole(
-    @Body() userRoleData: CreateUserRoleDto,
-  ): Promise<UserRoleResponseDTO> {
-    return await this.userService.addUserRole(userRoleData);
-  }
-
-  @AuthenticatedUser({ isAdmin: true })
-  @ApiTags('roles')
-  @ApiOperation({ summary: 'Update existing user role' })
-  @ApiParam({ name: 'userRoleId', required: true, type: 'integer' })
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description: 'Returns the updated user role',
-    type: UserRoleResponseDTO,
-  })
-  @ApiResponse({
-    status: HttpStatus.BAD_REQUEST,
-    description: 'Role already exists',
-  })
-  @Put('roles/:userRoleId')
-  public async updateUserRole(
-    @Param() params,
-    @Body() userRoleData: UpdateUserRoleDto,
-  ): Promise<UserRoleResponseDTO> {
-    return await this.userService.updateUserRole(
-      params.userRoleId,
-      userRoleData,
-    );
-  }
-
-  @AuthenticatedUser({ isAdmin: true })
-  @ApiTags('roles')
-  @ApiOperation({ summary: 'Delete existing user role' })
-  @ApiParam({ name: 'userRoleId', required: true, type: 'integer' })
-  // TODO: REFACTOR: rename to /users/roles/
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description: 'Returns the deleted role',
-    type: UserRoleResponseDTO,
-  })
-  @ApiResponse({
-    status: HttpStatus.NOT_FOUND,
-    description: 'No role found',
-  })
-  @Delete('roles/:userRoleId')
-  public async deleteUserRole(
-    @Param('userRoleId', ParseIntPipe)
-    userRoleId: number,
-  ): Promise<UserRoleResponseDTO> {
-    return await this.userService.deleteUserRole(userRoleId);
-  }
-
   @AuthenticatedUser({ isOrganizationAdmin: true })
-  @ApiTags('users')
   @ApiOperation({ summary: 'Get all users' })
   @ApiResponse({
     status: HttpStatus.OK,
@@ -154,7 +64,6 @@ export class UserController {
 
   // TODO: define response type, this cannot use an interface though
   @AuthenticatedUser({ isOrganizationAdmin: true })
-  @ApiTags('users')
   @ApiOperation({ summary: '[EXTERNALLY USED] Sign-up new user' })
   @ApiResponse({
     status: HttpStatus.NO_CONTENT,
@@ -167,7 +76,6 @@ export class UserController {
   }
 
   @Throttle(THROTTLING_LIMIT_HIGH)
-  @ApiTags('users')
   @ApiOperation({ summary: '[EXTERNALLY USED] Log in existing user' })
   @ApiResponse({
     status: HttpStatus.CREATED,
@@ -210,7 +118,6 @@ export class UserController {
   }
 
   @AuthenticatedUser()
-  @ApiTags('users')
   @ApiOperation({ summary: 'Log out existing user' })
   @Post('users/logout')
   public async logout(@Res() res): Promise<UserRO> {
@@ -233,7 +140,6 @@ export class UserController {
 
   @Throttle(THROTTLING_LIMIT_HIGH)
   @AuthenticatedUser()
-  @ApiTags('users')
   @ApiOperation({ summary: 'Change password of logged in user' })
   // TODO: Change this in to a PATCH request
   @Post('users/password')
@@ -249,7 +155,6 @@ export class UserController {
   }
 
   @AuthenticatedUser({ isAdmin: true })
-  @ApiTags('users')
   @ApiOperation({ summary: 'Delete user by userId' })
   @ApiResponse({
     status: HttpStatus.OK,
@@ -266,7 +171,6 @@ export class UserController {
   }
 
   @AuthenticatedUser()
-  @ApiTags('users')
   @ApiOperation({ summary: 'Get current user' })
   @Get('users/current')
   @ApiResponse({
@@ -286,7 +190,6 @@ export class UserController {
   }
 
   @AuthenticatedUser({ permissions: [PermissionEnum.AidWorkerProgramUPDATE] })
-  @ApiTags('users')
   @ApiOperation({
     summary:
       'Search, across all programs, for users who are already part of a program or who can be added to a program, based on their username or a substring of their username.' +
@@ -309,155 +212,7 @@ export class UserController {
     return await this.userService.findUsersByName(username);
   }
 
-  @AuthenticatedUser({ isAdmin: true })
-  @ApiTags('users/assignments')
-  @ApiOperation({ summary: 'Get roles for given user program assignment' })
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description: 'Returns program assignment including roles and scope',
-    type: AssignmentResponseDTO,
-  })
-  @ApiResponse({
-    status: HttpStatus.NOT_FOUND,
-    description: 'No roles found for user',
-  })
-  @Get('programs/:programId/users/:userId')
-  public async getAidworkerProgramAssignment(
-    @Param('programId', ParseIntPipe)
-    programId: number,
-
-    @Param('userId', ParseIntPipe)
-    userId: number,
-  ): Promise<AssignmentResponseDTO> {
-    return await this.userService.getAidworkerProgramAssignment(
-      programId,
-      userId,
-    );
-  }
-
-  @AuthenticatedUser({ permissions: [PermissionEnum.AidWorkerProgramUPDATE] })
-  @ApiTags('users/assignments')
-  @ApiOperation({
-    summary: 'Create or OVERWRITE program assignment including roles and scope',
-  })
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description:
-      'Returns the created or overwritten program assignment including roles and scope',
-    type: AssignmentResponseDTO,
-  })
-  @ApiResponse({
-    status: HttpStatus.NOT_FOUND,
-    description: 'User, program or role(s) not found',
-  })
-  @Put('programs/:programId/users/:userId')
-  public async assignAidworkerToProgram(
-    @Param('programId', ParseIntPipe)
-    programId: number,
-
-    @Param('userId', ParseIntPipe)
-    userIdToUpdate: number,
-
-    @Body()
-    assignAidworkerToProgram: CreateProgramAssignmentDto,
-
-    @Req() req: ScopedUserRequest,
-  ): Promise<AssignmentResponseDTO> {
-    throwIfSelfUpdate(req, userIdToUpdate);
-    return await this.userService.assignAidworkerToProgram(
-      programId,
-      userIdToUpdate,
-      assignAidworkerToProgram,
-    );
-  }
-
-  @AuthenticatedUser({ permissions: [PermissionEnum.AidWorkerProgramUPDATE] })
-  @ApiTags('users/assignments')
-  @ApiOperation({
-    summary:
-      'Update existing program assignment with new roles (UNION of existing and new roles) and/or overwrite scope',
-  })
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description: 'Returns program assignment with all roles and scope',
-    type: AssignmentResponseDTO,
-  })
-  @ApiResponse({
-    status: HttpStatus.NOT_FOUND,
-    description: 'User, program or role(s) not found',
-  })
-  @Patch('programs/:programId/users/:userId')
-  public async updateAidworkerProgramAssignment(
-    @Param('programId', ParseIntPipe)
-    programId: number,
-    @Param('userId', ParseIntPipe)
-    userIdToUpdate: number,
-    @Body() assignAidworkerToProgram: UpdateProgramAssignmentDto,
-    @Req() req: ScopedUserRequest,
-  ): Promise<AssignmentResponseDTO> {
-    throwIfSelfUpdate(req, userIdToUpdate);
-    return await this.userService.updateAidworkerProgramAssignment(
-      programId,
-      userIdToUpdate,
-      assignAidworkerToProgram,
-    );
-  }
-
-  @AuthenticatedUser({ permissions: [PermissionEnum.AidWorkerProgramUPDATE] })
-  @ApiTags('users/assignments')
-  @ApiOperation({
-    summary:
-      'Remove roles from program-assignment (pass roles to delete in body) or remove assignment (no body)',
-  })
-  @ApiBody({ type: DeleteProgramAssignmentDto, required: false })
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description:
-      'Returns the program assignment with the remaining roles or nothing (if program assignment removed)',
-    type: AssignmentResponseDTO,
-  })
-  @ApiResponse({
-    status: HttpStatus.NOT_FOUND,
-    description: 'User, program, role(s) or assignment not found',
-  })
-  @Delete('programs/:programId/users/:userId')
-  public async deleteAidworkerRolesOrAssignment(
-    @Param('programId', ParseIntPipe)
-    programId: number,
-
-    @Param('userId', ParseIntPipe)
-    userIdToUpdate: number,
-
-    @Req() req: ScopedUserRequest,
-    @Body() deleteProgramAssignment?: DeleteProgramAssignmentDto,
-  ): Promise<AssignmentResponseDTO | void> {
-    throwIfSelfUpdate(req, userIdToUpdate);
-
-    return await this.userService.deleteAidworkerRolesOrAssignment({
-      programId,
-      userId: userIdToUpdate,
-      roleNamesToDelete: deleteProgramAssignment?.rolesToDelete,
-    });
-  }
-
-  @AuthenticatedUser({ permissions: [PermissionEnum.AidWorkerProgramREAD] })
-  @ApiTags('users/assignments')
-  @ApiOperation({ summary: 'Get all users by programId' })
-  @ApiResponse({
-    status: HttpStatus.OK,
-    description: 'Returns a list of users assigned to a program',
-    type: [GetUserReponseDto],
-  })
-  @Get('programs/:programId/users')
-  public async getUsersInProgram(
-    @Param('programId', ParseIntPipe)
-    programId: number,
-  ): Promise<GetUserReponseDto[]> {
-    return await this.userService.getUsersInProgram(programId);
-  }
-
   @AuthenticatedUser({ isOrganizationAdmin: true })
-  @ApiTags('users')
   @ApiOperation({ summary: 'Reset user password without current password' })
   @ApiResponse({
     status: HttpStatus.NO_CONTENT,
@@ -475,7 +230,6 @@ export class UserController {
 
   // Make sure this endpoint is below users/password endpoint in this controller, to avoid endpoint confusion
   @AuthenticatedUser({ isAdmin: true })
-  @ApiTags('users')
   @ApiOperation({
     summary:
       'Update user properties (currently isOrganizationAdmin and displayName)',
