+++ services/121-service/src/user/user.module.ts
@@ -3,10 +3,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { ProgramEntity } from '@121-service/src/programs/entities/program.entity';
 import { ProgramAidworkerAssignmentEntity } from '@121-service/src/programs/entities/program-aidworker.entity';
+import { RolesController } from '@121-service/src/user/controllers/roles.controller';
+import { UserController } from '@121-service/src/user/controllers/user.controller';
+import { UserAssignmentsController } from '@121-service/src/user/controllers/user-assignments.controller';
 import { PermissionEntity } from '@121-service/src/user/entities/permissions.entity';
 import { UserEntity } from '@121-service/src/user/entities/user.entity';
 import { UserRoleEntity } from '@121-service/src/user/entities/user-role.entity';
-import { UserController } from '@121-service/src/user/user.controller';
 import { UserService } from '@121-service/src/user/user.service';
 import { UserEmailsModule } from '@121-service/src/user/user-emails/user-emails.module';
 
@@ -22,7 +24,7 @@ import { UserEmailsModule } from '@121-service/src/user/user-emails/user-emails.
     UserEmailsModule,
   ],
   providers: [UserService],
-  controllers: [UserController],
+  controllers: [UserController, UserAssignmentsController, RolesController],
   exports: [UserService],
 })
 export class UserModule {}
